Fix: Issue caused by splitting code in diff files
